install: start a test suite for GisScribe
glibtests.m4 and glib-tap.mk are from GLib Git, with a couple of path
changes in the latter, marked with comments.

This version of tap-test comes from ostree. Unlike the upstream version,
it runs tests in a subdirectory of /var/tmp rather than in the
source/build dir. This is of interest if one is building in a VirtualBox
shared folder, for instance, which doesn't support cool features like
hardlinks. (Without this, the libglnx tests, coming soon, fail.)